Remove unecessary helpers ...
... and move other helper functions to the `helpers.go` file
