+++ pkg/cliplugins/workspace/plugin/helpers.go
@@ -17,41 +17,113 @@ limitations under the License.
 package plugin
 
 import (
+	"context"
+	"fmt"
+	"net/url"
+	"path"
 	"strings"
 
-	"k8s.io/client-go/tools/clientcmd/api"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/kcp-dev/kcp/pkg/virtual/workspaces/registry"
+	virtualcommandoptions "github.com/kcp-dev/kcp/cmd/virtual-workspaces/options"
+	tenancyhelpers "github.com/kcp-dev/kcp/pkg/apis/tenancy/v1alpha1/helper"
+	tenancyv1beta1 "github.com/kcp-dev/kcp/pkg/apis/tenancy/v1beta1"
+	tenancyclient "github.com/kcp-dev/kcp/pkg/client/clientset/versioned"
 )
 
-// prioritizedAuthInfo returns the first non-nil or non-empty AuthInfo it finds
-// from the ordred list passed in argument.
-// The first return paremeter is the index at which the returned AuthInfo was found.
-// -1 means none was found.
-func prioritizedAuthInfo(values ...*api.AuthInfo) (int, *api.AuthInfo) {
-	for i, value := range values {
-		if value == nil {
-			continue
-		}
-		value := *value
-		if value.Token != "" || value.TokenFile != "" || value.Password != "" || value.Username != "" ||
-			value.Exec != nil || value.AuthProvider != nil {
-			return i, &value
-		}
+// getWorkspaceFromInternalName retrieves the workspace with this internal name in the
+// user workspace directory, by requesting the `workspaces` virtual workspace.
+func getWorkspaceFromInternalName(ctx context.Context, workspaceInternalName string, tenancyClient tenancyclient.Interface) (*tenancyv1beta1.Workspace, error) {
+	if list, err := tenancyClient.TenancyV1beta1().Workspaces().List(ctx, metav1.ListOptions{
+		FieldSelector: "metadata.name=" + workspaceInternalName,
+	}); err != nil {
+		return nil, err
+	} else if list == nil || len(list.Items) == 0 {
+		return nil, fmt.Errorf("workspace %q is not found", workspaceInternalName)
+	} else if len(list.Items) > 1 {
+		return nil, fmt.Errorf("several workspaces with the same internal name : %q", workspaceInternalName)
+	} else {
+		return &list.Items[0], nil
 	}
-	return -1, api.NewAuthInfo()
 }
 
-// extractScopeAndName extracts the scope and workspace name from
-// workspace kubeconfig context names as they are set in workspace Kubeconfigs
-// retrieved from the virtual workspace kubeconfig sub-resource
-func extractScopeAndName(workspaceKey string) (string, string) {
-	if strings.HasPrefix(workspaceKey, registry.PersonalScope+"/") {
-		return registry.PersonalScope, strings.TrimPrefix(workspaceKey, registry.PersonalScope+"/")
-	} else if strings.HasPrefix(workspaceKey, registry.OrganizationScope+"/") {
-		return registry.OrganizationScope, strings.TrimPrefix(workspaceKey, registry.OrganizationScope+"/")
+// getWorkspaceAndBasePath gets the workspace name, org logical cluster name and the base URL for the current
+// workspace.
+func getWorkspaceAndBasePath(urlPath string) (orgClusterName, workspaceName, basePath string, err error) {
+	// get workspace from current server URL and check it point to an org or the root workspace
+	serverURL, err := url.Parse(urlPath)
+	if err != nil {
+		return "", "", "", err
 	}
-	return "", workspaceKey
+
+	possiblePrefixes := []string{
+		"/clusters/",
+		path.Join(virtualcommandoptions.DefaultRootPathPrefix, "workspaces") + "/",
+	}
+
+	var clusterName string
+	for _, prefix := range possiblePrefixes {
+		clusterIndex := strings.Index(serverURL.Path, prefix)
+		if clusterIndex < 0 {
+			continue
+		}
+		clusterName = strings.SplitN(serverURL.Path[clusterIndex+len(prefix):], "/", 2)[0]
+		basePath = serverURL.Path[:clusterIndex]
+	}
+
+	if clusterName == "" {
+		return "", "", basePath, fmt.Errorf("current cluster URL %s is not pointing to a workspace", serverURL)
+	}
+
+	var org string
+	if clusterName == tenancyhelpers.RootCluster {
+		orgClusterName = ""
+		workspaceName = tenancyhelpers.RootCluster
+	} else if org, workspaceName, err = tenancyhelpers.ParseLogicalClusterName(clusterName); err != nil {
+		return "", "", "", fmt.Errorf("unable to parse cluster name %s", clusterName)
+	} else if org == "system:" {
+		return "", "", "", fmt.Errorf("no workspaces are accessible from %s", clusterName)
+	} else if org == tenancyhelpers.RootCluster {
+		orgClusterName = tenancyhelpers.RootCluster
+	} else {
+		orgClusterName, err = tenancyhelpers.ParentClusterName(clusterName)
+		if err != nil {
+			// should never happen
+			return "", "", "", fmt.Errorf("unable to derive parent cluster name for %s", clusterName)
+		}
+	}
+
+	return orgClusterName, workspaceName, basePath, nil
+}
+
+// upToOrg derives the org workspace cluster name to operate on,
+// from a given workspace logical cluster name.
+func upToOrg(orgClusterName, workspaceName string, always bool) string {
+
+	if orgClusterName == "" && workspaceName == tenancyhelpers.RootCluster {
+		return tenancyhelpers.RootCluster
+	}
+
+	if orgClusterName == tenancyhelpers.RootCluster && !always {
+		return tenancyhelpers.EncodeOrganizationAndClusterWorkspace(tenancyhelpers.RootCluster, workspaceName)
+	}
+
+	return orgClusterName
+}
+
+func outputCurrentWorkspaceMessage(orgName, workspacePrettyName, workspaceName string, opts *Options) error {
+	if workspaceName != "" {
+		message := fmt.Sprintf("Current workspace is %q", workspacePrettyName)
+		if workspaceName != workspacePrettyName {
+			message = fmt.Sprintf("%s (an alias for %q)", message, workspaceName)
+		}
+		if orgName != "" {
+			message = fmt.Sprintf("%s in organization %q", message, orgName)
+		}
+		err := write(opts, fmt.Sprintf("%s.\n", message))
+		return err
+	}
+	return nil
 }
 
 func write(opts *Options, str string) error {
